commit my code pawn and request loan nft

+++ lib/data/response/pawn/offer_sent/offer_sent_detail_crypto_response.dart
@@ -110,7 +110,7 @@ class OfferSentDetailCryptoResponse extends Equatable {
   int? bcOfferId;
 
   @JsonKey(name: 'liquidationThreshold')
-  int? liquidationThreshold;
+  double? liquidationThreshold;
 
   OfferSentDetailCryptoResponse(
     this.id,
